+++ crates/plotinator_mqtt/Cargo.toml
@@ -18,6 +18,10 @@ strum.workspace = true
 plotinator_macros.workspace = true
 anyhow.workspace = true
 
+[dev-dependencies]
+test_util.workspace = true
+testresult.workspace = true
+
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 rumqttc.workspace = true
